MailGenrator/requirements.txt
annotated-types==0.5.0
anyio==3.7.1
certifi==2023.7.22
click==8.1.7
colorama==0.4.6
distro==1.8.0
exceptiongroup==1.1.3
fastapi==0.103.1
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
openai==1.3.7
pydantic==2.3.0
pydantic_core==2.6.3
python-dotenv==1.0.0
PyYAML==6.0.1
sniffio==1.3.0
starlette==0.27.0
tqdm==4.66.1
typing_extensions==4.7.1
uvicorn==0.23.2
requests==2.31.0
